+++ react/ProductSearchContextProvider.js
@@ -6,7 +6,6 @@ import { Helmet, withRuntimeContext } from 'render'
 import DataLayerApolloWrapper from './components/DataLayerApolloWrapper'
 import searchQuery from './queries/searchQuery.gql'
 import {
-  canonicalPathFromParams,
   createInitialMap,
   SORT_OPTIONS,
 } from './utils/search'
@@ -35,6 +34,14 @@ class ProductSearchContextProvider extends Component {
       order: PropTypes.oneOf(SORT_OPTIONS.map(o => o.value)),
       priceRange: PropTypes.string,
     }),
+    /** Custom query `query` param */
+    queryField: PropTypes.string,
+    /** Custom query `map` param */
+    mapField: PropTypes.string,
+    /** Custom query `rest` param */
+    restField: PropTypes.string,
+    /** Custom query `orderBy` param */
+    orderByField: PropTypes.string,
     /** Current extension point name */
     nextTreePath: PropTypes.string,
     /** Component to be rendered */
@@ -152,16 +159,21 @@ class ProductSearchContextProvider extends Component {
     const from = (page - 1) * maxItemsPerPage
     const to = from + maxItemsPerPage - 1
 
+    const includeFacets = (map, query) => !!(map && map.length > 0 && query && query.length > 0)
+
+    const query = Object.values(params)
+      .filter(s => s.length > 0)
+      .join('/')
+
     const defaultSearch = {
-      query: Object.values(params)
-        .filter(s => s.length > 0)
-        .join('/'),
+      query,
       map,
       rest,
       orderBy,
       priceRange,
       from,
       to,
+      withFacets: includeFacets(map, query),
     }
 
     const customSearch = {
@@ -172,6 +184,7 @@ class ProductSearchContextProvider extends Component {
       priceRange,
       from,
       to,
+      withFacets: includeFacets(mapField, queryField),
     }
 
     return (
@@ -196,12 +209,6 @@ class ProductSearchContextProvider extends Component {
               loading={loading}
             >
               <Helmet>
-                {params && (
-                  <link
-                    rel="canonical"
-                    href={`https://${window.__hostname__}${canonicalPathFromParams(params)}`}
-                  />
-                )}
                 {titleTag && <title>{titleTag}</title>}
                 {metaTagDescription && (
                   <meta name="description" content={metaTagDescription} />