+++ react/utils/canonical.js
@@ -1,0 +1,7 @@
+import RouteParser from 'route-parser'
+
+const canonicalPathFromParams = (canonicalTemplate, params) => {
+  return new RouteParser(canonicalTemplate).reverse(params)
+}
+
+export default canonicalPathFromParams